==> notebooks/3_Merge_Given_Dataset.ipynb <==
# Merge Merchant and Consumer Features

This notebook merges consumer and merchant data into the transaction records to create a complete dataset. Also converts postcode to SA2 codes to prepare for merging with external datasets in the next step.

## Convert Postcode to SA2

# baseline imports
from pathlib import Path
import pandas as pd
import numpy as np
import pyarrow.parquet as pq
import pyarrow as pa

RAW_EXT = Path("../data/raw/external_dataset")
CLE = Path("../data/cleaned")
CUR = Path("../data/curated")

consumer_df = pd.read_parquet(CUR / "consumer_features.parquet")
merchant_df = pd.read_parquet(CUR / "merchant_features.parquet")
transactions = pd.read_parquet(CLE / "transaction.parquet")

# Postcode to SA2 concordance table
col_types = {"POSTCODE": str, "SA2_CODE_2021":str, "RATIO_FROM_TO": float}
postcode_sa2_df = pd.read_excel(RAW_EXT / "CG_POSTCODE_2021_SA2_2021.xlsx", converters=col_types)

# Find how many rows has a missing postcode
null_ids = int(postcode_sa2_df["POSTCODE"].isna().sum())
null_ids

# Drop rows with missing postcodes 
postcode_sa2_df = postcode_sa2_df.dropna(subset=["POSTCODE"])  

# For each postcode, keep the SA2 with the largest area share
sa2_ratio = (postcode_sa2_df.groupby("POSTCODE")["RATIO_FROM_TO"]
                                .max()
                                .reset_index())

# Merge back to get the corresponding SA2 code & name
preferred_sa2 = postcode_sa2_df.merge(sa2_ratio,on=["POSTCODE", "RATIO_FROM_TO"])

# Keep only needed columns, rename for consistency
sa2_poa_codes = preferred_sa2.rename(
    {
        "POSTCODE": "postcode",
        "SA2_CODE_2021": "SA2_code",
        "SA2_NAME_2021": "SA2_name"
    },
    axis=1
)[["postcode", "SA2_code", "SA2_name"]]

sa2_poa_codes.head(10)

## Add SA2 to consumer dataset

print(consumer_df.shape)

consumer_sa2 = consumer_df.merge(
    sa2_poa_codes,
    on="postcode",
    how="left"
)

# Count rows where SA2 is null
null_count = consumer_sa2['SA2_code'].isna().sum()
print(f"Number of rows with null SA2: {null_count}")

Some postcodes are only for PO Boxes and Large Volume Receiver. LVR is when many large businesses, government departments and other institutions receiving high volumes of mail had their own postcode. We decided to exclude these from our analysis.

Data is taken from - https://www.matthewproctor.com/australian_postcodes#downloadlinks

# Define PO Box / LVR postcode ranges
po_box_ranges = [
    ("1000", "1999"), ("0200", "0299"), ("8000", "8999"),
    ("9000", "9999"), ("5800", "5999"), ("6800", "6999"),
    ("7800", "7999"), ("0900", "0999")
]

# Flag PO Box postcodes
def is_po_box(pc):
    return any(lower <= pc <= upper for lower, upper in po_box_ranges)

consumer_sa2["is_po_box"] = consumer_sa2["postcode"].apply(is_po_box)

# Split missing SA2 into PO/LVR vs not
no_sa2_mask = consumer_sa2["SA2_code"].isna()
no_sa2_but_po = consumer_sa2[ no_sa2_mask &  consumer_sa2["is_po_box"]]
no_sa2_not_po = consumer_sa2[ no_sa2_mask & ~consumer_sa2["is_po_box"]]

print("Counts:", {
    "no_sa2_but_po": len(no_sa2_but_po),
    "no_sa2_not_po": len(no_sa2_not_po)
})
print("\nPO/LVR postcodes with no SA2:", np.sort(no_sa2_but_po["postcode"].unique()))
print("\nNon-PO postcodes with no SA2:", np.sort(no_sa2_not_po["postcode"].unique()))

# Total before filtering
total_consumers = consumer_sa2["consumer_id"].nunique()

# Count removals
removed_po_box = consumer_sa2[consumer_sa2["is_po_box"]]
removed_unknown_sa2 = consumer_sa2[(consumer_sa2["SA2_code"].isna())]

n_removed_po_box = removed_po_box["consumer_id"].nunique()
n_removed_unknown = removed_unknown_sa2["consumer_id"].nunique()

# Combined removal (avoid double-counting)
removed_combined = consumer_sa2[
    consumer_sa2["is_po_box"] | consumer_sa2["SA2_code"].isna()]
n_removed_total = removed_combined["consumer_id"].nunique()

pct_removed_total = n_removed_total / total_consumers * 100

print(f"Total consumers: {total_consumers}")
print(f"Removed (PO Box/LVR): {n_removed_po_box}")
print(f"Removed (Unknown SA2): {n_removed_unknown}")
print(f"Total removed: {n_removed_total} ({pct_removed_total:.2f}%)")
print(f"Remaining consumers: {total_consumers - n_removed_total}")

We decided to remove all consumers whose postcodes fall within **PO box** or **LVR (Large Volume Receiver)** ranges. These addresses typically do not represent the consumer’s place of residence and therefore cannot be reliably mapped to a meaningful SA2 region. 

For the small number of remaining consumers with unknown or unmapped SA2 codes, we also excluded them from further analysis. Attempting to impute SA2 values for these consumers could introduce noise or bias, particularly because our downstream modelling relies heavily on external datasets (GCP, WPP, APRA-SEIFA) that are SA2-based. By filtering out these records early, we ensure that the features used in our models accurately reflect the socioeconomic and demographic context of the consumers we are analysing, leading to more robust and interpretable results.

# Drop rows
print(f"Shape before drop: {consumer_sa2.shape}")
consumer_sa2_cleaned = consumer_sa2[~(consumer_sa2["SA2_code"].isna())]
consumer_sa2_cleaned = consumer_sa2_cleaned.drop(columns=["is_po_box"])

# Count rows where SA2 is null
null_count = consumer_sa2_cleaned['SA2_code'].isna().sum()
print(f"Number of rows with null SA2: {null_count}")
print(f"Shape after drop: {consumer_sa2_cleaned.shape}")

consumer_sa2_cleaned.head(5)

consumer_sa2_cleaned.to_csv(CUR / "consumer_sa2.csv")

## Merge consumer, merchant and transactions

# Rename columns to avoid confusion
consumer_sa2 = consumer_sa2.rename(columns={"fraud_prob_avg": "consumer_fraud_prob_avg"})
consumer_sa2 = consumer_sa2.rename(columns={"name": "consumer_name"})

merchant_df = merchant_df.rename(columns={"name": "merchant_name"})

# Join consumer features into transactions
tx_cons = transactions.merge(
    consumer_sa2,
    on="user_id",
    how="left"
)

# Join merchant features into transactions
tx_full = tx_cons.merge(
    merchant_df,
    on="merchant_abn",
    how="left"
)

tx_full.head()

tx_full.shape

# count NaN values per column
nan_counts = tx_full.isna().sum()
nan_counts

# Save to curated folder
tx_full.to_parquet(CUR / "all_given_data.parquet", index=False)

## Summary

**all_given_data:**

-----Transaction-----
- **user_id** : used to merge consumer features
- **merchant_abn** : used to merge merchant features
- **dollar_value**
- **order_id**
- **order_datetime**

-----Consumer features----- 
- **consumer_id** → unique identifier for the consumer  
- **transaction_count** → total number of transactions made  
- **total_spend** → total dollar value spent across all transactions  
- **unique_merchants** → number of distinct merchants purchased from  
- **avg_spend** → average spend per transaction  
- **consumer_fraud_prob_avg** → average fraud probability across transactions  
- **consumer_name** → consumer’s name  
- **address** → consumer’s address  
- **state** → state of residence  
- **postcode** → postcode of residence  
- **gender** → gender of the consumer 
- **SA2_code** → SA2 code of the consumer 
- **SA2_name** → SA2 name of the consumer 

-----Merchant features----- 
- **merchant_name** → merchant’s name  
- **tags** → descriptive tags associated with the merchant  
- **categories** → category labels for the merchant  
- **type** → merchant type classification  
- **take_rate** → BNPL commission rate (%)  
- **order_count** → total number of transactions with this merchant     
- **total_sales** → total revenue generated from transactions  
- **avg_order_value** → average value per transaction    
- **unique_consumers** → number of distinct customers served  
- **repeat_consumers** → number of customers with 2+ purchases  
- **repurchase_rate** → proportion of customers who are repeat buyers
