==> notebooks/9.1_Segmentation_Manual.ipynb <==
# Segmentation: Manual Ranking

## Import Libraries and Starting Pyspark

from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.sql.types import *
import pandas as pd
import requests
import re
import numpy as npv
from pyspark.sql.window import Window

# Create a spark session
spark = (
    SparkSession.builder.appName("Manual Ranking Segmentation")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.driver.memory", "9g") 
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .config("spark.network.timeout", "600s")
    .getOrCreate()
)

## Load Data

merchant = spark.read.parquet(str("../data/curated/merchant_features.parquet"))
merchant_fraud = spark.read.parquet(str("../data/curated/merchant_imputed.parquet"))

merchant_df = merchant.toPandas()

merchant_df.columns.tolist()

merchant_df.categories.value_counts()

def clean_text_any(x):
    if x is None or (isinstance(x, float) and pd.isna(x)):
        s = ""
    # if it's already a list -> join
    elif isinstance(x, (list, tuple, set)):
        s = " ".join(map(str, x))
    else:
        s = str(x)

    s = s.lower()
    s = re.sub(r"[^a-z0-9\s]", " ", s)   # keep letters/numbers/spaces
    s = re.sub(r"\s+", " ", s).strip()
    return s

# clean → tokens → remove stopwords → join back
stop_words = set(["is","a","the","for","and","to","of","this","except","in","on","at","with","by"])
merchant_df["str_categories"]   = merchant_df["categories"].apply(clean_text_any)
merchant_df["tokens"]           = merchant_df["str_categories"].str.split()
merchant_df["clean_categories"] = merchant_df["tokens"].apply(lambda xs: [w for w in xs if w not in stop_words])
merchant_df["clean_categories_str"] = merchant_df["clean_categories"].apply(lambda xs: " ".join(xs))

merchant_df["clean_categories_str"].value_counts(dropna=False).head(30)

# Manually categorise each category
segments = [
    "Entertainment & Media", "Entertainment & Media", "Technology", "Beauty", 
    "Miscellaneous", "Office & Home Supplies", "Technology", 
    "Entertainment & Media", "Office & Home Supplies", "Technology", 
    "Beauty", "Miscellaneous", "Entertainment & Media", 
    "Beauty", "Entertainment & Media", "Office & Home Supplies", 
    "Office & Home Supplies", "Miscellaneous", "Miscellaneous", 
    "Entertainment & Media", "Office & Home Supplies", 
    "Office & Home Supplies", "Technology", "Entertainment & Media", "Beauty"
]

segments_dict = {list(merchant_df.clean_categories_str.value_counts().index)[i]: segments[i] for i in range(len(segments))}

segments_dict

merchant_df['segment']  = merchant_df['clean_categories_str'].map(lambda x: segments_dict[x])

merchant_df.segment.value_counts()

merchant_df[['merchant_abn', 'segment', 'categories']].to_parquet("../data/curated/merchant_segment.parquet")

## Ranking based on each segment

initial_ranking = spark.read.parquet("../data/curated/merchant_ranking.parquet")
merchant_segment = spark.read.parquet("../data/curated/merchant_segment.parquet")

# Join each merchant in initial ranking by segment
merchant_segment_ranking = initial_ranking.join(merchant_segment, how='left', on='merchant_abn')

segments = [
    "Entertainment & Media",
    "Office & Home Supplies",
    "Miscellaneous",
    "Beauty",
    "Technology"
]

for segment in segments:
    print(segment)
    segment_ranking = merchant_segment_ranking.filter(F.col('segment') == segment)
    segment_ranking.orderBy(F.col('final_score').desc()).show(10)

## Save top 10 merchants to parquet

candidate = (
    merchant_segment_ranking
    .filter(F.col("segment").isin(segments))
    .filter(F.col("final_score").isNotNull())
)

w = Window.partitionBy("segment").orderBy(
    F.col("final_score").desc(),
    F.col("merchant_abn").asc()
)

top10_per_segment = (
    candidate
    .withColumn("rank", F.row_number().over(w))
    .filter(F.col("rank") <= 10)
)

for seg in segments:
    (
        top10_per_segment
        .filter(F.col("segment") == seg)
        .orderBy("rank")
        .write.mode("overwrite")
        .parquet(f"../data/curated/top10_manual_{re.sub(r'[^A-Za-z0-9]+','_', seg.lower())}.parquet")
    )


## Justification and Insights

If a merchant has high revenue + high repeat + low fraud, it is considered as top ranked. 
High fraud (either total $ fraud or high avg consumer fraud) pulls scores down, even with big revenue.

1) Entertainment & Media
- Pattern: high repurchase (0.94–1.00) and strong revenue_norm; fraud_sum_norm is often high (0.69–0.83), but the inverse terms partially offset it.
- Takeaway: This segment earns high ranks when repeat is strong. Where repeat collapses, even big revenue can’t carry them.
- Action:
    - Grow: high-repurchase, moderate-fraud merchants (e.g., music/digital goods).
    - Watch: accounts with low repeat despite high revenue—probe CX, refunds, or category churn.
2) Office & Home Supplies
- Pattern: wide spread. Some have excellent repeat (~1.0) and big revenue, but others show extreme avg consumer fraud (e.g., 206–241), dragging scores.
- Takeaway: Fraud quality (avg_cons_fraud) is the decisive separator within this segment.
- Action:
    - De-risk: merchants with avg_cons_fraud > ~60 → tighten KYC/velocity rules, lower limits, step-up auth.
    - Upsell: the near-1.0 repurchase cohort (tents/furniture) with loyalty offers; they have sticky demand.
3) Miscellaneous
- Pattern: consistently high repurchase (~0.9–1.0) + high revenue_norm; fraud generally mid, so inverses help → top scores (0.76–0.70).
- Takeaway: Solid “core” segment for profitable growth; fraud stays in check.
- Action:
    - Scale: marketing credit lines/promos to top repeat merchants.
    - Monitor: sub-categories with avg_cons_fraud > ~70.
4) Beauty
- Pattern: high repurchase and revenue_norm at the top (scores ~0.78), but many mid-ranked merchants have moderate fraud and middling repeat (0.35–0.66).
- Takeaway: Beauty is attractive where loyalty is strong; otherwise, scores flatten.
- Action:
    - Grow: high-repeat beauty merchants (loyalty-driven SKUs).
    - Fix: low-repeat sub-categories (shoes/watches) with targeted incentives or better UX.
5) Technology
- Pattern: revenue often strong, but avg_cons_fraud spikes (e.g., 87–234) + variable repeat → scores cluster around 0.54–0.75.
- Takeaway: Fraud quality & low repeat are the blockers. Tech looks good only where repeat stays high.
- Action:
    - De-risk: telecom/computers with avg_cons_fraud > ~80 (hard flags).
    - Keep/grow: software/services with sticky subscriptions (high repeat).
